services/web-api/requirements.txt
fastapi==0.109.*
uvicorn[standard]==0.27.*
sqlalchemy==2.0.*
psycopg2-binary==2.9.*
pydantic==2.6.*
google-cloud-secret-manager==2.*
opentelemetry-api==1.*
opentelemetry-sdk==1.*
opentelemetry-exporter-gcp-trace==1.*
opentelemetry-instrumentation-fastapi==0.*
opentelemetry-propagator-gcp==1.*
